Move volume thresholding tools to their own module (#387)
They don't really belong to the skinseg module
