+++ OpenLIFULib/OpenLIFULib/session.py
@@ -8,7 +8,7 @@ from slicer import (
 )
 from slicer.parameterNodeWrapper import parameterPack
 from OpenLIFULib.util import get_openlifu_data_parameter_node
-from OpenLIFULib.skinseg import load_volume_and_threshold_background
+from OpenLIFULib.volume_thresholding import load_volume_and_threshold_background
 from OpenLIFULib.lazyimport import openlifu_lz
 from OpenLIFULib.parameter_node_utils import SlicerOpenLIFUSessionWrapper, SlicerOpenLIFUPhotoscanWrapper
 from OpenLIFULib.targets import (
